messy code, can read in genome xml files
The real change was adding code to SharpNEATLib that allows for random
generation of genomes - with or without recurrence. On top of that, I am
writing out sampled data points. Next step is to load those objects into
the CPPN.js library and parse them into CPPNs. Then test the CPPNs.
Then, the final step is to create tests on top of the CPPN library to
validate things didn't break after changes are made in the future. Clean
this mess up.