WOOT. Working CPPNs. According to the manual tests.
Set up all the infrastructure for loading in and decoding a CPPN object
from XML. Also, set up the testing. Requires a certain XML file format.
I have included 2 examples in Data. Both are not recurrent - haven't
tested recurrence yet. You load in a file, and the console will output
the expected and the actual. The first 8 digits check out - but the last
8 are different. This is due to doubles/floats in my SharpNEATLib CPPN
execution. Next phase will be adding in checks to make sure things are
formatting, as well as reasonable defaults for missing information. And
then I'll make some D3 examples of CPPN stuff.