Switched cppnjs to now be a component/nodejs module.
This is a pretty cool change as it now can be used browser side with all
that weird loadLibraryFile syntax.

The cppn library should be buildable and used standalone in the
browser/node. There's also a concept for "extras" where certain classes
can be extended.

This marks version 0.2.0 of the cppn library. I won't push the npm until
the transition is complete for at least cppnjs and neatjs.